% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{importAllMdf}
\alias{importAllMdf}
\title{Importing multiple file types and storing it into a list of data frames}
\usage{
importAllMdf(path = NULL)
}
\arguments{
\item{path}{the directory in which the files are stored.
       If path is not given, interactive dialog box will be used to select directory}
}
\value{
a single list of dataframes containing all the files imported and stored as dataframes.
}
\description{
importAllMdf function loads various data file types in a selected directory to separate dataframes,
and stores them as a list. Data frames can be accessed as list elements by using "listname$filename" or "listname[]".
}

