% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{importAllSdf}
\alias{importAllSdf}
\title{Importing multiple file types and storing it into a Single R Dataframe}
\usage{
importAllSdf(path = NULL)
}
\arguments{
\item{path}{the directory in which the files are stored.
       If path is not given, interactive dialog box will be used to select directory}
}
\value{
a single dataframe containing all the files imported and rbinded into one.
}
\description{
importAllSdf function loads various data file types in a selected directory to dataframes,
combines all the data frames and stores it as a single data frame.
Note that all the files to be loaded should have the same number of columns for rbind to work
}

