% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInDataset.R
\name{read_in_dataset}
\alias{read_in_dataset}
\title{Load a dataset into R}
\usage{
read_in_dataset(file, method = c("data.table", "RData", "feather", "rds"))
}
\arguments{
\item{file}{file to be read into R. Must be tabular data}

\item{method}{the library to load the file into R}
}
\description{
Instead of relying on remembering and loading multiple types of data loading
libraries, this function is a wrapper that contains several.
}
